requirements.txt
streamlit
matplotlib
pandas
numpy
seaborn
fpdf
Pillow